Add project method to return defunct collectors + their aliases
This helps with cases where a collector has been re-named, but the
project created a symlink between the old name and the new one.
By default, we remove defunct collectors from the collector list,
but it is still possible that a user might use the old name in a
query.

Now each project finder can define its own set of defunct
collector names and any possible aliases that can be used in
their place when servicing a query -- that should allow a query
for a collector that we've removed (e.g. route-views2.saopaulo) to
still return a usable result (albeit referring to files under
ix-br2.gru) rather than simply failing with "collector not found".