Disable all the route-views2.saopaulo redirects I just added
Turns out that since ix-br2.gru isn't linked from the main
routeviews URL, our scraper doesn't recognize it (or any of
the other new collectors) as a collector.

So trying to replace route-views2.saopaulo right now doesn't make
sense but we will need to do so whenever routeviews gets around to
updating their front page.

Note: they have an API for getting the collector set and we may
need to start using that in the future to avoid this type of
discrepancy.