Update collector row even if latest file is missing
The columns in the collector table that track last crawl times and
the most recent file were not being updated because each
scrape attempt didn't manage to get *all* of the files we were
expecting.

Given that our RV archive is always going to be 1 update behind
due to our mirroring process, that's not really something that
should be treated as a "fatal" error.

Instead, we will continue to retry if the latest file is missing,
but if it is not there after we run out of retries then we won't
throw an error.